fork init
