+++ src/ffi/utils.rs
@@ -1,0 +1,14 @@
+use std::slice::from_raw_parts;
+
+pub unsafe fn get_str<'a>(s: *const u16) -> &'a [u16] {
+    unsafe fn strlen(s: *const u16) -> usize {
+        let mut i = 0;
+        while *s.add(i) != 0 {
+            i += 1;
+        }
+        i + 1
+    }
+
+    let len = strlen(s);
+    from_raw_parts(s, len)
+}
